chore(tests_integration): move http client test util
commit-id:2d145538
