+++ crates/starknet_http_server/src/lib.rs
@@ -2,3 +2,5 @@ pub mod communication;
 pub mod config;
 pub mod errors;
 pub mod http_server;
+#[cfg(feature = "testing")]
+pub mod test_utils;
